base: clean up and refactor containers.
The containers no longer force the user to use the ptrdiff_t type for
size information. For convenience, size() now always returns an int,
and the size_s() method offers a way to handle arrays with more than
2 billion elements. Internally, we still use ptrdiff_t, though.

Since so much code had to be changed, I took the opportunity to get
rid of capitalised accessors for classes that are lowercase.

+++ src/platform/sdl/sdlapp.cpp
@@ -198,13 +198,14 @@ void SdlAppDisplay::Disable()
 }
 
 #if USE_SDL
-ptrdiff_t SceneDisplay::GetPhysicalCount()
+int SceneDisplay::GetPhysicalCount()
 {
-    return (ptrdiff_t)SDL_GetNumVideoDisplays();
+    return SDL_GetNumVideoDisplays();
 }
-const char* SceneDisplay::GetPhysicalName(ptrdiff_t index)
+
+const char* SceneDisplay::GetPhysicalName(int index)
 {
-    return SDL_GetDisplayName((int)index);
+    return SDL_GetDisplayName(index);
 }
 #elif USE_OLD_SDL && USE_D3D9
 //  Not implemented
